Shrink memory usage by reducing max_depth
This commit changes max_depth field type from int (4 bytes) to uint8_t
(1 byte) in rb_t structure, saving 3 bytes per tree instance while
maintaining correctness since the theoretical maximum tree depth is 121
for 64-bit systems, well within uint8_t range (0-255).

+++ rbtree.c
@@ -325,8 +325,8 @@ void rb_insert(rb_t *tree, rb_node_t *node)
     fix_extra_red(stack, stacksz);
 
     /* Update the maximum depth of the tree if necessary. */
-    if (stacksz > (unsigned) tree->max_depth)
-        tree->max_depth = (int) stacksz;
+    if (stacksz > tree->max_depth)
+        tree->max_depth = (uint8_t) stacksz;
 
     /* Ensure the root is correctly updated after potential rotations. */
     tree->root = stack[0];
